P2-stuff/requirements.txt
argparse
grpcio
grpcio-tools
protobuf
google-cloud-storage
google-cloud-firestore
jwcrypto